Added modified versions of project website files
Since the old project repository had pages for the game and high scores, I
decided to copy them over to ease the setup of this repository's website.

It's very possible that additional changes will need to be made to make the
repository website work, but I won't know until I can test it.